+++ changes.md
@@ -7,6 +7,7 @@
 * Added ramp_frame.py example
 * Added RampGreebled
   * Added Portal Greebled Ramp Example
+* Integrated hinges into the Portal
 
 ## 0.1.1
 * As predicted fixed as issue with ramp inheriting width from frame in unexpected and exciting ways.